refactor: 주차 정보 API Mapper 구조로 리팩토링 #232
